feature: merge analysis_helpers.py into helpers.py
+ remove unused imports and add some annotations to tests
